refactor: extract GitHubExportProvider+IssueCreation (#639 slice G, FINAL)
Seventh and final slice of #639. Moves the issue-creation + retry
cluster into a dedicated extension file. When this merges, #639
(GitHubExportProvider decomposition epic, 821 lines original) is
fully sliced and can be closed.

Moved (5 members):
- func export(issues:session:configuration:) — already public, unchanged
- func makeURLRequest(issue:session:configuration:exportFingerprint:) — already internal test seam, unchanged
- static let retryContext (with 5-line doc comment) — stays private in new file (same-file callers)
- createWithRetry(issue:fingerprint:request:configuration:successfulCount:) — stays private (same-file caller: export)
- attemptCreate(_:configuration:) — stays private (same-file caller: createWithRetry)

Stored-property widenings:
- logger: private let → let (used by export.info calls + createWithRetry)
- retryConfiguration: private let → let (used by createWithRetry)

Post-slice state: all 5 GitHubExportProvider actor stored properties
now have module-internal visibility (session/receiptStore/retryConfiguration/
logger/annotationRenderer). Base class body is now init-only.

Line-count delta:
- GitHubExportProvider.swift: 230 → 27 (-203 lines; only stored props + init remain)
- GitHubExportProvider+IssueCreation.swift (new): 188 lines
- xcodeproj: +4

Test coverage: GitHubExportProviderTests passes unchanged.

Cumulative decomposition of #639 (7 slices A/B/C/D/E/F/G):
- GitHubExportProvider.swift: 821 → 27 (-794 lines, -97%)
- 7 extension files added: WireTypes (88), Networking (77), Duplicates
  (155), IssueBody (172), RepositoryDiscovery (87), Preflight (47),
  IssueCreation (188) = 814 lines
- Net repo change: +20 lines for extension boilerplate

Closes #888.
Closes #639.
